Modification des configs
